test.Runner: Allow to build self-contained tests
For our upcoming preloader, we're going to preload using macros to
substitute the values wherever we use the preloader.

This means, that we can't simply test different preloaders in one run
and have to separate them into different files, because as soon as we
mark something for preloading it will affect all other uses of the
preloader.

So we now have a small macro function within the Runner class, which
simply adds a main() function so that the test case can be directly
compiled and executed separately from the main tests.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>