preloader: Introduce a proxy class for results
So far we went through reflection whenever we wanted to get one of the
results of the preloader, which is quite bad in terms of performance.

Now we're copying the values from the PreloaderSteps.results field to
the PreloaderProxy.results field, because we can directly do a lookup on
PreloaderProxy while we can't look up PreloaderSteps without reflection.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>