AnimationTest: Add tests for unnecessary updates
We already have a few tests for that, but we really want to make sure
that whenever an animation frame stays the same we don't get a redraw.

I've added two test cases, one for when an animation only consists of a
single frame and another one if a specific key consists of a single
frame.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>