SpriteTest: Add test to check for update overlap
This test is obviously just verifying the current implementation,
however what we want to do is to cut down the update rectangle within
every sprite, so that we don't need to update things like characters
around the sprite with full alpha.

So when we do that implementation all we need to do is making very small
modifications to this test and we're set.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>