image: Implement as a real class instead of alias
The Image class now is actually an extension (but not a subclass) of
Raster and also shares the same interface.

What I have in mind here is that the Image class also has a dirty
rectangle, which includes _only_ the changes that have happened since
the last update.

Having such a dirty rect allows us to skip not only transparent
characters but it should also perform better with very large sprites and
particularly animations that have very few changes.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>