sprite: Hook in new Image implementation
We now are rendering by subdividing the areas of every sprite, which
saves on the redraw area of them. This also makes a few hacks from
Animation obsolete, where we had manually set is_dirty once a new frame
was in position.

I haven't yet benchmarked the results between this implementation and
the one before, but depending on the results, we might want to use it
only for larger sprites.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>