curses: Improve terminal rendering
So far we had only a very rudimentary implementation of the curses
rendering, although it's still very rudimentary when it comes to drawing
optimization, because we could make shortcuts from the current cursor
position using CSI that move relatively.

But in order to do that, we need to have knowledge about the current
cursor position at all times, which by this commit we do have now.

We now also have a complete implementation of clear().

Signed-off-by: aszlig <aszlig@redmoonstudios.org>