Change license to GNU AGPL version 3 (only)
So first of all, why did I choose the AGPL?

The world as of today has mostly accepted games being closed source,
which is a shame and I'd really like to change that in my little
library, which is why I choose a very strong copyleft here.

In the first place, this software was licensed under the MIT/Expat
license, because I was feeling that the software is not substantial
enough to choose a strong copyleft.

Nowadays however, I really don't care whether it's substantial or not,
because if I'd be the single one user of this software for eternity then
let it be so :-)

Needless to say, that with this license change I do intend to prevent
this software from being used in any proprietary applications, including
my own ones (in the unlikely event that this should happen one day).

Second, why only version 3?

While I don't believe that the FSF is going to introduce a newer license
that I don't like (the license itself states that it will be "similar in
spirit"), my approach is however to first read the license before
blindly accepting it.

This is a bit like a terms of service agreements of US companies, where
a lot of them state "will change without notice, you need to check back
here regularily". Even if it would be a company which I'd trust (haha),
I really can't be sure at which point they're going nuts.

I've changed the indenting of the license text a bit, but otherwise the
full license is included unaltered in COPYING.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>