image: Fix dirty_rect coordinates for resizing
This is a similar off-by-one error that we already had in
4d32688b2271b19946f4fc7c2f958a2e31fa79a4.

And I fucked up badly with this, because I was counting from zero, but
that won't work with those rectangles, so let's illustrate.

Imagine you have an Image with the dimensions (8, 4), the full rect
would be (0, 0, 8, 4):

  012345678   So let's say we want this   012345678
  1.------.   rect to be increased in     1.------.
  2|      |   height by 2.                2|      |
  3|      |                               3|      |
  4`------'   This will look like this:   4|------|
                                          5|      |
                                          6`------'

So we now have the original rectangle, plus the actual expansion
rectangle which has the dimensions (0, 4, 2, 2).

The bug here was that instead of (0, 4, 2, 2) I created the rectangle
with (0, 3, 2, 2), which is clearly wrong, because we don't need to
0-index anything here because the rectangle's height is _already_ its
bottom Y position.

In addition to fixing the bug, I've also fixed the test and added a few
more animation tests, because that's where rendering bugs are most
likely to show up, should we have another such fuckup.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>