image: Don't set dirty when growing
Growing the sprite will fill everything with transparency, so we really
don't need to expand the dirty_rect, because it will be set dirty as
soon as some of the raster cells are modified.

However, we still need to set it dirty whenever we remove a row or
column, because we need to update the region which then no longer
exists.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>