image: Always use Symbol(0) as the default value
The frontends have so far relied on the surface image being filled with
spaces, but handling Symbol(0) should be really a matter of the
frontend, so we can keep Image without an option to specify different
default values.

Having a way to specify default values would mean that we can't make
certain assumptions like "if you grow the image there will be only
transparency", which gets in our way of optimizing the rendering chain.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>