image: Implement bounding box for dirty_rect
Using Raster is very heavyweight, especially when it comes to resizing,
because an array has to be spliced around.

We now have a bounding box which serves as a reference for the dirty
rectangle, so whenever a redraw has finished, the dirty rectangle is set
to that bounding box again.

The bounding box itself tries very hard to trim off any transparent
characters and looking at it in debug mode seems that we have
accomplished that.

Our underlying Raster type is now also just written to whenever it's
_really_ necessary and the raster only grows but doesn't shrink, so we
avoid unneccessary garbage collection runs.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>