PVector: Fix return of normalize() method
For most cases, the normalize() method returns a new copy of PVector,
but if the length of the PVector is 0.0 or 1.0 the normalize method
returns a reference to the initial value.

So altering the result of normalize() would sometimes alter the original
PVector, which is really not as intended.

Apart from adding a test case, I've also uninlined the method and fixed
a few calls to static normalize() which I've worked around in an earlier
commit (6ecd0edebaefbf434429325597edce81fd19ec4e).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>