object: Base everything on PVector and thus floats
One of the major mistakes we did here, was to make our coordinates of
type integer. Floats come with their own issues, especially when it
comes to different platforms, but we already use floats to calculate
most of our algorithms so we can't really avoid them altogether anyway.

With integers we constantly needed to convert between floats and ints,
especially when it comes to something like the td value in conjuction
with incremential movement.

Another downside of using integers there is that we can't really work
with vectors in conjuction with sprites/objects except if we assign
temporary variables which we then round to integers.

So this should now get rid of these restrictions, but the implementation
is still pretty basic, because we have a few new issues here:

  * A lot more garbage is now generated from PVector (de-)allocations.
  * Unlinke plain integers, PVector types can be passed by reference and
    are mutable.

We can however mitigate both if we use memory pooling for PVector and
make sure that PVector operations are immutable at all times.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>