object: Add a rotate_around() method
While we have our perfectly valid rotate() function for PVector, it's
not enough to just rotate by an angle in a circular shape.

We're rendering character-based, so the height of the characters is
twice the size of the width. So what we need to do is to actually rotate
around the object in an elliptic shape.

By scaling the vectors prior to rotation we get exactly that effect.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>