test.hxml: Use neko target instead of --macro
It's not possible to run tests involving macros if we run the whole test
suite within macro context. We already run the tests for their
respective targets in default.nix, so we should do so here as well.

In addition to just running the tests, we also remove the build result
so that we don't leave garbage around.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>