Add a new Renewable interface along with macros
We want to have memory pooling, but in order to implement that, we first
need to have a way to reinitialize objects that are no longer in use.

In theory this would be easy by just implementing an interface and write
a renew() method for each classe.

Unfortunately, this will only work in a very much limited way, because
while we can override methods from the super class we can't override it
with a different type signature.

Also, most of the time would have to duplicate the constructor code for
every renew() method.

Fortunately, macros make this a bit less annoying, so whenever a class
implements hase.iface.Renewable, an autoBuild macro creates renew
methods for every class and subclass. We do need to avoid using the same
names, so instead of generating renew() methods, we create __renew_N()
methods, where N is a suffix that's incremented for every class.

However, we still can't just use obj.renew() now, because of that name
mangling, so we now have a static method called Renew.object(), which
itself is a macro function, which resolves the right __renew_N() method
for the given object.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>