utils: Add a Pool class for memory pooling
It builds upon our Renewable interface we have implemented recently and
involves a bit of macro magic again.

There are basically two (macro) functions, which are Pool.alloc() and
Pool.free().

The former either creates a new instance (and thus expects to be passed
the constructor arguments) or if the pool has unused objects it returns
a renewed instance of it. The alloc() function *only* wants constructor
arguments, because the actual type is inferred from the calling context.

The latter just adds the object to the pool if it doesn't yet exist in
the unused range.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>