Improve shielding off macro functionality
This is particularly important when we use @:build and @:autoBuild,
because if we access such an interface or class while within macro
context, we get "You cannot use @:build inside a macro" as an error.

The error is basically a safety shield because if it wouldn't be there
we would run into an infinite recursion during macro evaluation.

So let's put a few #if !macro ... #end around macro functionality so
that we're not accidentally tripping on this.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>