pool: Fix pos information for alloc() and free()
We want alloc() and free() to use the position where it was called
instead of the position of the macro definition, so that it is clear
where the error is coming from. If we want to know to which macro
function this leads up we can still see that when looking at the line
where it was called.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>