pool: Allow to pass explicit type info to alloc()
While using the inferred type is sufficient in most cases, sometimes
it's useful to directly pass the type we want to have to the alloc
function.

We detect that by checking whether the first argument is a Class and if
that's the case we're resolving the type. If it fails to resolve, we eat
another argument, resolve it and pass it to the class that has been
resolved. The class along with the type param is then resolved again.

Unfortunately we can't simply use something like Pool.alloc(Foo<Int>),
because the parser rejects <> within method calls, so it's
Pool.alloc(Foo, Int) for now until we come up with a better way.

One downside to this approach is that if we have a class with multiple
type params, we can't resolve that when given explicitly.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>