Add a new Pooling interface for use with classes
So far we've had hase.utils.Pool, which pretty much relies on the
destination object/class to implement hase.iface.Renewable, which in
turn generates classes for Renew.object() to work.

We also have Pool.alloc() and Pool.free() which rely on Renewable to be
available for the target object/class.

All this is quite okay to implement for abstract types, but if we want
that on classes with inheritance, we need to write a lot of the same
code again and again for every single children of the superclass.

With hase.iface.Pooling this is no longer the case and all this can be
simply done by adding "implements hase.iface.Pooling" to the topmost
superclass.

This implementation currently has the limitation that it can't handle
type parameters very well, so that's where we need to improve on.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>