object: Use a List for the children attribute
We're reordering children several times, for example during changes of
the Z axis or when adding/removing subchilds. An array has the
disadvantage that we have a significant amount of overhead because the
array needs to be rebuilt from scratch whereas a linked list only needs
to change the links between the nodes.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>