surface: Use custom data structure for sprite list
We now have hase.ds.LinkedList, which is a singly linked list that's
optimized for our use case with sprites, so inserting items as well as
sorting them is cheap.

Calculating the length however would be quite expensive because we'd
need to recurse the whole list, so the data structure lacks things such
as .length or array indexing.

The merge sort algorithm is a cleaned up version from haxe.ds.ListSort
which in turn is from [1].

[1]: https://www.chiark.greenend.org.uk/~sgtatham/algorithms/listsort.html

Signed-off-by: aszlig <aszlig@redmoonstudios.org>