pool: Handle type parameters
Now it gets very ugly, because we can't use type parameters in static
variables by just put a @:generic and be done.

The reason this doesn't work is because specific types of the generics
are created during object construction and not on access.

We could have worked around this by creating another singleton class
which contains all of the pooled elements, but we'd still need to cast
around the container elements in the singleton class.

While keeping this in mind, I decided that if we're going the ugly
route, let's do the unsafe casting within PoolForType_N instead of
adding another global.

So the to-be-recycled objects now are monomorphic and also are casted
around to be bend into what the compiler infers for the first
instantiation. This of course has the downside that if we use pooling
for object that implement vastly different functionality depending on
its type, we're getting undefined behaviour.

Fortunately, these kinds of things are very rare and even if it's the
case, we can still implement manual pooling for these classes.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>