Use memory pooling wherever currently possible
Most of these are rather messy, because we can't override operators like
"+=" within abstracts. So we need to do workarounds like:

var tmp = valtochange + newval;
valtochange.free();
valtochange = tmp;

This especially becomes quite braindamaged in hase.display.Motion during
calculation of the follower forces.

Right now however, we don't have a really nice method to hide this
behind some macro magic, so we're left off doing it manually. At least
until we have our one-macro-fits-all sledgehammer.

For things such as LinkedListNode, I had to make the class public
because this is currently not possible very easily to access the type
from within macro context, see HaxeFoundation/haxe#3589 for more on
that.

After looking at the memory debugger of Firefox and Chromium, we still
have a few GC pauses because of some pesky arrays that are still
unpooled but it got a LOT better and for a very complex rendering scene
the average framerate (with profiling enabled) is around 55 FPS whereas
before we had around 11 FPS.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>