renewable: Allow to use metadata conditions
This allows to use @:new_only and @:renew_only metadata to annotate the
constructor function with things that either only are for construction
(new) or for reconstruction (renew).

It also somewhat obsoletes the ability to define a renew() method, but
the latter might still be useful if the renew() method is vastly
different than new() and would otherwise clutter up the new() method too
much.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>