image: Improve map_() and merge_raster() methods
The map_() method is better off inlined so that the function we pass
there is actually expanded into the code rather than passed as a
function at runtime.

In a similar vein the merge_raster() method so far uses bind() to pass
in the set() method, which not only is bad for debugging but also bad
for rendering performance.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>