pool: Add autofree() helper method
Since 5bd0221060fa4fa2dcfd7eac22c7ecc874d27d5f we now have a lot of ugly
code, where we add temporary variables just to free objects to the pool
that are intermediately allocated.

With this commit, we now have Pool.autofree(), which accepts any
expression (and of course blocks are allowed) and frees *any*
allocations to the pool once leaving the autofree() context.

One of the example would be the following:

  var foo = ((new PVector(1.0, 2.0) + new PVector(2.0, 4.0)) * 3.0).x;

This would allocate 4 intermediate objects and *not* free it to the
pool, so without autofree() we'd need to free() them manually like this:

  tmp1 = new PVector(1.0, 2.0);
  tmp2 = new PVector(2.0, 4.0);
  tmp3 = tmp1 + tmp2;
  tmp4 = tmp3 * 3.0;
  var foo = tmp4.x;
  tmp4.free();
  tmp3.free();
  tmp2.free();
  tmp1.free();

With Pool.autofree() all we now have to do is the following:

  var foo = Pool.autofree(
    ((new PVector(1.0, 2.0) + new PVector(2.0, 4.0)) * 3.0).x
  );

Of course, this will only work once the Disposable types for PVector are
in place.

As hinted already, there are two new types (or better: aliases) for
pooling: Allocated and Disposable. Both of them are in hase.mem.Types
and they are currently only for autofree to work correctly.

The difference between the two is that Allocated is returned whenever we
do a *direct* allocation and Disposable is only meant for functions that
indirectly allocate poolable objects.

One of the reasons is that if we have some static variable which is
allocated via Pool.alloc() and we use autofree(), the static variable is
freed onto the Pool and might potentially be invalidated for other
functions.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>