pvector/rect: Mark types as Disposable/Allocated
While this isn't really a big deal for ordinary objects, it really does
make a difference for abstract types, because we really need to
distinguish whether the returned value was directly (Allocated) or
indirectly (Disposable) allocated and we can't do that for the
underlying type of the abstract.

The reason for the latter is that in abstracts we handle pooling by
ourselves and don't have a builder interface for that which automates
this.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>