path: Add iterator() and partition()
This is mostly to make it easier to get equally distributed points along
the path. The iterator() method is just for convenience and it's
re-using the iterator() method of Array.

The partition() method iterates over the whole path using the specified
increments and it's basically a convenience wrapper for pos_at().

Signed-off-by: aszlig <aszlig@nix.build>