Convert into using Nix Flakes
This is essentially the same as we had in the default.nix so far, but
makes it easier to re-use in other projects because we get a known
working nixpkgs revision and we don't need to rely on import-from-
derivation.

Additionally, we also get Hydra jobsets without the need to create a
dedicated release.nix.

Signed-off-by: aszlig <aszlig@nix.build>