Refactor example build expression into Flake
The example already targets cpp, js and neko and most of my other
projects using Hase use a similar expression to build. In order to
facilitate code reuse, it makes sense to generalise and expose the
build expression so that it can be used by other Flakes.

With this we now also build the example as part of the Flake.

Signed-off-by: aszlig <aszlig@nix.build>