Explicitly provide Hydra jobs
So far we only exposed all the attributes in check, which so far were
only the unit tests. However, for Hydra it also makes sense to build the
examples and even allow to run the examples in-browser.

Additionally, this lets us get rid of a few systems that we don't want
to have Hydra builds for, mainly everything but x86_64-linux right now.

Signed-off-by: aszlig <aszlig@nix.build>